@@ -1,6 +1,3 @@
-@using Microsoft.AspNetCore.Identity
-@using PostgreSqlDotnetCore
+﻿@using PostgreSqlDotnetCore
 @using PostgreSqlDotnetCore.Models
-@using PostgreSqlDotnetCore.Models.AccountViewModels
-@using PostgreSqlDotnetCore.Models.ManageViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
